feat(ActionWorker, ActionNode, Timeout): moved wrapped_action_work to ActioNodeWorker, added functional timeout ability, tested
