+++ beams/behavior_tree/ActionNode.py
@@ -1,69 +1,15 @@
 import atexit
 import logging
 import os
-import time
-from multiprocessing import Event, Queue, Value
-from typing import Callable
+from multiprocessing import Event
 
 import py_trees
 
-from beams.behavior_tree.ActionWorker import ActionWorker
+from beams.behavior_tree.ActionWorker import ActionWorker, wrapped_action_work  # the latter is grabbed as a pas through
 from beams.behavior_tree.VolatileStatus import VolatileStatus
-from beams.typing_helper import (ActionNodeWorkFunction, ActionNodeWorkLoop,
-                                 Evaluatable)
+from beams.typing_helper import ActionNodeWorkLoop, Evaluatable
 
 logger = logging.getLogger(__name__)
-
-
-def wrapped_action_work(loop_period_sec: float = 0.1):
-    def action_worker_work_function_generator(func: ActionNodeWorkFunction) -> ActionNodeWorkLoop:
-        def work_wrapper(
-            do_work: Value,
-            name: str,
-            work_gate: Event,
-            volatile_status: VolatileStatus,
-            completion_condition: Evaluatable,
-            log_queue: Queue,
-            log_configurer: Callable) -> None:
-            """
-            Wrap self.work_func, and set up logging / status communication
-            InterProcess Communication performed by shared memory objects:
-            - volatile status
-            - logging queue
-
-            Runs a persistent while loop, in which the work func is called repeatedly
-            """
-            log_configurer(log_queue)
-            while (do_work.value):
-                logger.debug(f"WAITING FOR INIT from node: {name}")
-                work_gate.wait()
-                work_gate.clear()
-
-                # Set to running
-                volatile_status.set_value(py_trees.common.Status.RUNNING)
-                while not completion_condition():
-                    logger.debug(f"CALLING CAGET FROM from node ({name})")
-                    try:
-                        status = func(completion_condition)
-                    except Exception as ex:
-                        volatile_status.set_value(py_trees.common.Status.FAILURE)
-                        logger.error(f"Work function failed, setting node ({name}) "
-                                     f"as FAILED. ({ex})")
-                        break
-
-                    volatile_status.set_value(status)
-                    logger.debug(f"Setting node ({name}): {volatile_status.get_value()}")
-                    time.sleep(loop_period_sec)
-
-                # one last check
-                if completion_condition():
-                    volatile_status.set_value(py_trees.common.Status.SUCCESS)
-                else:
-                    volatile_status.set_value(py_trees.common.Status.FAILURE)
-
-                logger.debug(f"Worker for node ({name}) completed.")
-        return work_wrapper
-    return action_worker_work_function_generator
 
 
 class ActionNode(py_trees.behaviour.Behaviour):
